[Codegen] Extend ConvertAccGemmToGeem to handle pooling with optional fill op.
The revision fills the `tensor.empty` op with a neutral value, based on the combiner op. It extends the pass to handle pooling ops properly.

Furthermore, it tracks the SSA chain to determine if the accumulated tensor is ReadOnly or not, instead of having a list. It is okay to not perform the rewrite if the destination tensor has writing access, because the bufferization can re-use the destination buffer without creating a stack buffer.

Signed-off-by: hanhanW <hanhan0912@gmail.com>