Carry encoding in the preferred storage type of a hoistable type
This PR makes sure that the encoding attribute is included in the preferred storage type of a HoistableTensorType. Without this we can get a tensor.bitcast on an encoded type returning a non-encoded type, which will result in a compilation failure when converting into stream.

This PR adds a convertForBitcast interface method to the SerializableAttr interface for encodings to specify how the types should be converted on a bitcast operation. This adds generality as different encodings can provide different implementations, allows them to perform correctness checks and would even allow propagating encodings further on the bitcasted type if needed in the future.

Signed-off-by: hanhanW <hanhan0912@gmail.com>